Minor cleanup

+++ modules/hlp.c
@@ -466,7 +466,7 @@ static void do_bplustree(deark *c, lctx *d, de_int64 pos1, de_int64 len)
 	for(pass=1; pass<=2; pass++) {
 		de_int64 curr_page;
 
-		de_memset(page_seen, 0, d->bpt.num_pages);
+		de_memset(page_seen, 0, (size_t)d->bpt.num_pages);
 
 		de_dbg(c, "pass %d\n", pass);
 		de_dbg_indent(c, 1);
